# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M International AI Quote Agent operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rfqs:
    get:
      operationId: listRfqs
      tags: [operations]
      summary: List RFQs
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/RequestType"
        - $ref: "#/components/parameters/Priority"
        - $ref: "#/components/parameters/Status"
      responses:
        "200":
          description: RFQ list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rfq"
  /rfqs/{rfqId}:
    get:
      operationId: getRfq
      tags: [operations]
      summary: Get an RFQ
      parameters:
        - $ref: "#/components/parameters/RfqId"
      responses:
        "200":
          description: RFQ details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfqDetail"
        "404":
          description: RFQ not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ai-review/{rfqId}/action:
    post:
      operationId: reviewAi
      tags: [operations]
      summary: Apply a review queue action
      parameters:
        - $ref: "#/components/parameters/RfqId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiReviewAction"
      responses:
        "200":
          description: Updated review item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReview"
        "404":
          description: RFQ not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    RfqId:
      name: rfqId
      in: path
      required: true
      schema:
        type: integer
...
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
    Source:
      name: source
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/RfqSource"
    RequestType:
      name: requestType
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/RequestType"
    Priority:
      name: priority
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Priority"
    Status:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/RfqStatus"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    RfqSource:
      type: string
      enum: [EMAIL, WEBSITE, PHONE, PORTAL]
    RequestType:
      type: string
      enum: [PARTS_EXCHANGE, NEW_PART_PURCHASE, REPAIR, OVERHAUL, INSPECTION, UNKNOWN]
    RfqStatus:
      type: string
      enum: [NEW, ANALYZING, UNDER_REVIEW, PROCESSING, NEEDS_HUMAN_REVIEW, COMPLETED, CLOSED]
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH, URGENT]
...
    Rfq:
      type: object
      properties:
        id:
          type: integer
        rfqNumber:
          type: string
        customer:
          type: string
        customerCompany:
          type: string
        partNumber:
          type: string
        source:
          $ref: "#/components/schemas/RfqSource"
        requestType:
          $ref: "#/components/schemas/RequestType"
        confidence:
          type: integer
        priority:
          $ref: "#/components/schemas/Priority"
        status:
          $ref: "#/components/schemas/RfqStatus"
        createdAt:
          type: string
        age:
          type: string
      required: [id, rfqNumber, customer, customerCompany, partNumber, source, requestType, confidence, priority, status, createdAt, age]
    RfqDetail:
      allOf:
        - $ref: "#/components/schemas/Rfq"
        - type: object
          properties:
            description:
              type: string
            quantity:
              type: integer
            aircraft:
              type: string
            notes:
              type: string
            emailSubject:
              type: string
            sender:
              type: string
            sourceEmail:
              oneOf:
                - $ref: "#/components/schemas/EmailDetail"
                - type: "null"
            analysis:
              oneOf:
                - $ref: "#/components/schemas/AiAnalysis"
                - type: "null"
            reviewHistory:
              type: array
              items:
                $ref: "#/components/schemas/ReviewHistoryEntry"
          required: [description, quantity, aircraft, notes, emailSubject, sender]
    Email:
      type: object
      properties:
        id:
          type: integer
        sender:
          type: string
        senderEmail:
          type: string
        subject:
          type: string
        receivedAt:
          type: string
        aiStatus:
          type: string
          enum: [ANALYZED, PROCESSING, FAILED]
        requestType:
          $ref: "#/components/schemas/RequestType"
        confidence:
          type: integer
        status:
          type: string
          enum: [RFQ_CREATED, PENDING_REVIEW, IGNORED]
      required: [id, sender, senderEmail, subject, receivedAt, aiStatus, requestType, confidence, status]
    EmailClassification:
      type: string
      enum: [RFQ, NON_RFQ, SUPPLIER_RESPONSE, CUSTOMER_INQUIRY, INTERNAL, UNKNOWN]
    ProcessingStatus:
      type: string
      enum: [NEW, SYNCED, PROCESSING, PROCESSED, REVIEW_REQUIRED, FAILED, IGNORED]
    AttachmentProcessingStatus:
      type: string
      enum: [PENDING, PROCESSING, EXTRACTED, FAILED, UNSUPPORTED]
    EmailAttachment:
      type: object
      properties:
        id:
          type: integer
        emailId:
          type: integer
        fileName:
          type: string
        contentType:
          type: string
        fileSize:
          type: integer
        processingStatus:
          $ref: "#/components/schemas/AttachmentProcessingStatus"
        extractedText:
          type: ["string", "null"]
      required: [id, emailId, fileName, contentType, fileSize, processingStatus, extractedText]
    AiAnalysis:
      type: object
      properties:
        id:
          type: integer
        emailClassification:
          $ref: "#/components/schemas/EmailClassification"
        requestType:
          $ref: "#/components/schemas/RequestType"
        confidenceScore:
          type: number
        reasoningSummary:
          type: string
        requiresHumanReview:
          type: boolean
        developmentMode:
          type: boolean
        extractedData:
          type: object
          additionalProperties: true
      required: [id, emailClassification, requestType, confidenceScore, reasoningSummary, requiresHumanReview, developmentMode, extractedData]
    ReviewHistoryEntry:
      type: object
      properties:
        id:
          type: integer
        action:
          type: string
          enum: [APPROVED, REJECTED, RECLASSIFIED, EDITED]
        previousClassification:
          $ref: "#/components/schemas/RequestType"
        newClassification:
          $ref: "#/components/schemas/RequestType"
        notes:
          type: string
        createdAt:
          type: string
      required: [id, action, previousClassification, newClassification, notes, createdAt]
    EmailDetail:
      allOf:
        - $ref: "#/components/schemas/Email"
        - type: object
          properties:
            recipient:
              type: string
            cc:
              type: string
            bodyText:
              type: string
            bodyHtml:
              type: string
            processingStatus:
              $ref: "#/components/schemas/ProcessingStatus"
            emailClassification:
              $ref: "#/components/schemas/EmailClassification"
            attachments:
              type: array
              items:
                $ref: "#/components/schemas/EmailAttachment"
            analysis:
              oneOf:
                - $ref: "#/components/schemas/AiAnalysis"
                - type: "null"
            linkedRfq:
              oneOf:
                - $ref: "#/components/schemas/Rfq"
                - type: "null"
          required: [recipient, cc, bodyText, bodyHtml, processingStatus, emailClassification, attachments, analysis, linkedRfq]
...
    AiReview:
      type: object
      properties:
        id:
          type: integer
        rfqId:
          type: integer
        rfqNumber:
          type: string
        customer:
          type: string
        classification:
          $ref: "#/components/schemas/RequestType"
        confidence:
          type: integer
        reason:
          type: string
        status:
          type: string
          enum: [NEEDS_REVIEW, APPROVED, CHANGED]
      required: [id, rfqId, rfqNumber, customer, classification, confidence, reason, status]
    AiReviewAction:
      type: object
      properties:
        action:
          type: string
          enum: [REVIEW, APPROVE, CHANGE_CLASSIFICATION]
        requestType:
          $ref: "#/components/schemas/RequestType"
      required: [action]
...